fix(python): addresses are usize
Code and instance addresses cross into runtime symbols and back as
pointer-width integers rather than i64.

git-bug: a441135980fddcb958de71a1ffd3fa4325dc5afaa476f29db8e1ae41e63103aa

+++ crates/zyntax_builtins/src/dynamic.rs
@@ -508,7 +508,10 @@ pub(crate) fn declarations(policy: &Policy, list_type: TypeId) -> Vec<Decl> {
                 ),
                 vec![ret(call("zb_any_is", vec![a.e(), b.e()], boolean()))],
             ),
-            ret(eq(cast(raw_any(a.e()), i64()), cast(raw_any(b.e()), i64()))),
+            ret(eq(
+                cast(raw_any(a.e()), usize()),
+                cast(raw_any(b.e()), usize()),
+            )),
         ],
     ));
     // Identity: the singletons compare by category, heap objects by
@@ -527,8 +530,8 @@ pub(crate) fn declarations(policy: &Policy, list_type: TypeId) -> Vec<Decl> {
             when(
                 and(is(&ca, CUSTOM), is(&cb, CUSTOM)),
                 vec![ret(eq(
-                    cast(raw_any(a.e()), i64()),
-                    cast(raw_any(b.e()), i64()),
+                    cast(raw_any(a.e()), usize()),
+                    cast(raw_any(b.e()), usize()),
                 ))],
             ),
             when(
